tests/wpt/tests/content-security-policy/sandbox/window-reuse-unsandboxed.html

<!DOCTYPE html>
<html>

<head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>

<body>
    <script>
      var t = async_test("Window object should be reused");

      window.onmessage = t.step_func_done(function(e) {
        assert_equals(e.data, "test");
      });

      w = window.open("support/unsandboxed-post-property-to-opener.html","","width=400,height=400");
      w.testProperty = "test";
    </script>
</body>

</html>
